+++ activerecord/lib/active_record/relation.rb
@@ -41,11 +41,6 @@ module ActiveRecord
       reset
     end
 
-    def arel_attribute(name) # :nodoc:
-      table[name]
-    end
-    deprecate :arel_attribute
-
     def bind_attribute(name, value) # :nodoc:
       if reflection = klass._reflect_on_association(name)
         name = reflection.foreign_key